+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Sample median and sample variance
+- Sample median, variance, and stddev displayed
 
 ### Fixed